内嵌预导入 sql，配置文件移除 sql 资源路径
