+++ dao/init_dao.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	_ "github.com/go-sql-driver/mysql"
 	_ "modernc.org/sqlite"
-	"strings"
 	"xorm.io/xorm"
 )
 
@@ -43,48 +42,6 @@ func InitDatabase() error {
 	err = initDataImport()
 	if err != nil {
 		return fmt.Errorf("初始化数据导入失败 %v", err)
-	}
-	return nil
-}
-
-// initDataImport 初始化数据导入
-// 用于导入一些配置数据，以及图鉴网没有的数据
-func initDataImport() error {
-	// 初始化超级管理员信息
-	if len(config.AppConfig.Bot.Admin) > 0 {
-		var sql string
-		if config.AppConfig.DB.UseLocal {
-			sql = `INSERT OR IGNORE INTO admin (qq) values `
-		} else {
-			sql = `INSERT IGNORE INTO admin (qq) values `
-		}
-		qqValues := make([]string, 0, len(config.AppConfig.Bot.Admin))
-		for _, adminId := range config.AppConfig.Bot.Admin {
-			qqValues = append(qqValues, fmt.Sprintf("(%d)", adminId))
-		}
-		sql += strings.Join(qqValues, ",")
-		_, err := DB.Exec(sql)
-		if err != nil {
-			return fmt.Errorf("配置导入超级管理员信息出错 %v", err)
-		}
-	}
-
-	// 导入预配置数据
-	tableMap := map[string]interface{}{
-		"guest.sql":      database.Guest{},
-		"laboratory.sql": database.Laboratory{},
-	}
-	for file, table := range tableMap {
-		if total, err := DB.Count(table); err != nil {
-			return err
-		} else if total > 0 {
-			continue
-		} else {
-			_, err = DB.ImportFile(fmt.Sprintf("%s/%s", config.AppConfig.Resource.Sql, file))
-			if err != nil {
-				return err
-			}
-		}
 	}
 	return nil
 }
